requirements.txt
Flask
Pillow
sympy
pix2tex[gui]
latex2sympy2